requirements.txt
pandas==2.1.4
openpyxl==3.1.2
playwright==1.40.0
pywin32==306
Pillow==10.1.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.110.0